requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vcrpy==6.0.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.13.1

djangorestframework~=3.15.2